do not confuse a verify for a helper
verify_publish_command used verify_post_command as a helper, and that is
not good. helpers return status code but a verify not only returns a
status code but also echos an error to the user. we don't want to reuse
error messages, generally speaking. different spots that generate an
error most likely should have different error messages, because they are
most likely different errors.